% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d4storagehub4R.R
\docType{package}
\name{d4storagehub4R}
\alias{d4storagehub4R}
\alias{d4storagehub4R-package}
\title{R interface to the D4Science StorageHub/Workspace API}
\description{
R interface to the D4Science StorageHub/Workspace API
}
\details{
\tabular{ll}{
   Package: \tab d4storagehub4R\cr
   Type: \tab Package\cr
   Version: \tab 0.1\cr
   Date: \tab 2021-08-28\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
