% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoragehubManager.R
\docType{class}
\name{StoragehubManager}
\alias{StoragehubManager}
\title{StoragehubManager}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an StoragehubManager
}
\description{
StoragehubManager
}
\note{
Main user class to be used with \pkg{d4storagehub4R}
}
\section{Methods}{

\describe{
 \item{\code{new(url, token, logger)}}{
   This method is used to instantiate the \code{StoragehubManager}. By default,
   the url is inherited through D4Science Icproxy service.
   
   The token is mandatory in order to use Storage HUb API.
   
   The logger can be either NULL, "INFO" (with minimum logs), or "DEBUG" 
   (for complete curl http calls logs)
 }
 \item{\code{getToken()}}{
   Get Zenodo user token.
 }
}
}

\examples{
\dontrun{
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{manager}
\keyword{storagehub}
